[FIX] refactoring architecture
